+++ zephyr-mcumgr/src/lib.rs
@@ -4,7 +4,7 @@
 #![doc = include_str!("../README.md")]
 #![doc(issue_tracker_base_url = "https://github.com/Finomnis/zephyr-mcumgr-client/issues")]
 
-/// A high level client for Zephyr's MCUmgr SMP functionality
+/// A high-level client for Zephyr's MCUmgr SMP functionality
 pub mod client;
 pub use client::MCUmgrClient;