+++ CHANGELOG.md
@@ -7,11 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.7.1] - xxxx-xx-xx
 
-## Changes
+### Changes
 
 - Increase default timeout to `10000 ms`.
 
-## Fixes
+### Fixes
 
 - Firmware upload to MCUboot recovery mode failed with `MGMT_ERR_EOK`